move uart and spim modules to drivers/
