+++ platforms/allwinner-d1/boards/lichee-rv/src/main.rs
@@ -9,7 +9,8 @@ use drivers::{
     dmac::Dmac,
     plic::{Plic, Priority},
     timer::{Timer, TimerMode, TimerPrescaler, Timers},
-    uart::{kernel_uart, Uart},
+    uart::{kernel_uart, Uart, D1Uart},
+    spim::{kernel_spim1, SpiSenderClient, SpiSenderServer, SPI1_TX_DONE},
     Ram,
 };
 use kernel::{
@@ -20,10 +21,6 @@ use kernel::{
     },
     trace, Kernel, KernelSettings,
 };
-use spim::{kernel_spim1, SpiSenderClient, SpiSenderServer, SPI1_TX_DONE};
-use uart::D1Uart;
-mod spim;
-mod uart;
 
 const HEAP_SIZE: usize = 384 * 1024 * 1024;
