+++ platforms/allwinner-d1/drivers/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod dmac;
 pub mod plic;
+pub mod spim;
 pub mod timer;
 pub mod uart;
